requirements.txt
discord==2.3.2
discord.py==2.4.0
python-dotenv==1.0.1
asyncio==3.4.3
yt-dlp==2024.8.6
emoji==2.13.0